feat: automatically assign position and default name
The small blind is always p1, then comes p2 etc. If no explicit name is
specified for a player these values should be used.

Use plain objects instead of classes for Player because GH copilot
suggests it's easier to handle in React and I was struggling a bit with
dynamic state updates.