feat: don't make the buttons to inc/decrease initial stack 'tabbable'
This might be a mistake, but the same way that setting the dealer button
or deleting a player are currently not reachable via tab, it seems more
convenient like this. Tab will be used when entering data via a computer
and then typing in the correct number is probably fastest. The buttons
will be most useful on mobile devices.